Add base functionality;
This still hasn't been tested as I couldn't get the SMTP part working
locally. The flow is pretty much written out though. The send_email call
is commented out. I've tested the password reset validations, but the
UAA client call to reset the password still need to be written.

Also having some tests in place around this would be good. The current
test is a copy of the index.html test.