Update CSV parsing to correctly read file
The encoding format of the CSV files we use for checking valid government domains recently changed to UTF-8.  This changeset updates our CSV parsing to correctly open the file for parsing, and makes the code a bit more readable as well.

h/t to @ChrisMcGowan and @bengerman13 for tracking this down!